Refactor endpoint catch helpers
