+++ crates/rulemorph_endpoint/src/endpoint_engine.rs
@@ -28,6 +28,7 @@ use uuid::Uuid;
 
 use crate::ssrf::{ResolvedSsrfTarget, resolve_ssrf_target};
 
+mod catch;
 mod config;
 mod endpoint_rule;
 mod error;
@@ -38,6 +39,7 @@ mod rule_ref;
 mod trace_graph;
 mod validation;
 
+use self::catch::CatchSpec;
 pub use self::config::{ApiMode, EngineConfig, RequestContext};
 #[cfg(test)]
 use self::endpoint_rule::EndpointPath;
@@ -1302,45 +1304,6 @@ fn build_ssrf_audit_log(
         method: rule.request.method.clone(),
         url: redact_ssrf_url(url),
         reason: reason.to_string(),
-    }
-}
-
-#[derive(Debug)]
-struct CatchSpec(HashMap<String, String>);
-
-impl From<HashMap<String, String>> for CatchSpec {
-    fn from(value: HashMap<String, String>) -> Self {
-        CatchSpec(value)
-    }
-}
-
-impl CatchSpec {
-    fn match_target(&self, error: &EndpointError) -> Option<PathBuf> {
-        let map = &self.0;
-        if let Some(status) = error.status {
-            let key = status.to_string();
-            if let Some(value) = map.get(&key) {
-                return Some(PathBuf::from(value));
-            }
-            let pattern = if (400..500).contains(&status) {
-                "4xx"
-            } else if (500..600).contains(&status) {
-                "5xx"
-            } else {
-                ""
-            };
-            if !pattern.is_empty() {
-                if let Some(value) = map.get(pattern) {
-                    return Some(PathBuf::from(value));
-                }
-            }
-        }
-        if error.kind == EndpointErrorKind::Timeout {
-            if let Some(value) = map.get("timeout") {
-                return Some(PathBuf::from(value));
-            }
-        }
-        map.get("default").map(PathBuf::from)
     }
 }
